+++ src/audible/__init__.py
@@ -2,10 +2,9 @@
 
 from ._logging import log_helper
 from ._version import __version__
-from .auth import Authenticator, LoginAuthenticator, FileAuthenticator
+from .auth import Authenticator
 from .client import Client, AsyncClient
 
 __all__ = [
-    "__version__", "LoginAuthenticator", "FileAuthenticator", "Authenticator",
-    "log_helper", "Client", "AsyncClient"
+    "__version__", "Authenticator", "log_helper", "Client", "AsyncClient"
 ]